wdir_setup.sh: if pipeline files analysis*tab and config*yaml already exist, they are not overwritten

+++ wdir_setup.sh
@@ -11,8 +11,14 @@ fi
 mkdir -p ${wdir}/reference_transcripts
 
 ln -sf $(pwd -P)/quantizyme_model.3.snakefile ${wdir}
-cp cluster.yaml ${wdir}
-cp analysis.tab ${wdir}
+
+if [[ ! -f ${wdir}/cluster.yaml ]]; then
+    cp cluster.yaml ${wdir}
+fi
+
+if [[ ! -f ${wdir}/analysis.tab ]]; then
+    cp analysis.tab ${wdir}
+fi
 
 echo "pipeline_dir: \"`pwd`\"" | cat config.yaml - > ${wdir}/config.yaml
 
@@ -21,7 +27,10 @@ mkdir -p ${wdir}/match
 mkdir -p ${wdir}/read_datasets
 
 ln -sf $(pwd -P)/quantizyme_match.snakefile ${wdir}
-cp analysis.match.tab ${wdir}
+
+if [[ ! -f ${wdir}/analysis.match.tab ]]; then
+    cp analysis.match.tab ${wdir}
+fi
 
 echo "pipeline_dir: \"`pwd`\"" | cat config.match.yaml - > ${wdir}/config.match.yaml
 
@@ -30,9 +39,9 @@ tree ${wdir}
 
 echo ""
 echo "### MODEL workflow"
-echo "Please put your input reference transcripts in the `reference_transcripts` dir."
-echo "Distribution of lengths of reference transcripts will be located in the `reference_transcripts_length_distribution` folder."
+echo "Please put your input reference transcripts in the reference_transcripts dir."
+echo "Distribution of lengths of reference transcripts will be located in the reference_transcripts_length_distribution folder."
 echo ""
 echo "### MATCH workflow"
-echo "Please put your read datasets in fasta format in the `read_datasets` dir."
-echo "Results of the MATCH workflow will be located in the `match` folder."
+echo "Please put your read datasets in fasta format in the read_datasets dir."
+echo "Results of the MATCH workflow will be located in the match folder."
